+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
-* `3.0.40` Release - [3.0.3400](#30400)
+* `3.3.00` Release - [3.3.000](#33000)
+* `3.2.00` Release - [3.2.000](#32000)
+* `3.1.00` Release - [3.1.000](#31000)
+* `3.0.40` Release - [3.0.400](#30400)
 * `3.0.30` Release - [3.0.300](#30300)
 * `3.0.20` Release - [3.0.200](#30200)
 * `3.0.10` Release - [3.0.100](#30100)
@@ -43,10 +46,22 @@
 * `0.77.x` Releases - [0.77.0](#0770)
 
 
+## 3.3.000
+#### Feature
+* `EMP-18156` `EMP-18124` Allow developers to pass `materialProfile` & `customAdParams` 
+
+## 3.2.000
+#### Feature
+* `EMP-18131` pass number of ads & ad index during an ad break. This add two new parameters to `onWillPresentInterstitial` : noOfAds & adIndex
+
+## 3.1.000
+#### Feature
+* `EMP-17957` Allow app developers to pass `appName` to analytics
+
 ## 3.0.400
 #### Bug Fixes
 * `EMP-18019` Bug Fix :  SDK does not fire `onWillPresentInterstitial` event on pre roll ads
-
+* 
 ## 3.0.300
 #### Bug Fixes
 * `EMP-17986` Bug Fix :  Quartiles (ad Tracking Events) are sent twice from SDK